pyproject.toml
[project]
name = "resume-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.58.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
]
